Implemented unification for a single appbar
Implemented unification between MainAppBar and CourseAppBar components
