+++ components/layout/MainAppBar.tsx
@@ -1,23 +1,68 @@
 'use client'
 
-import * as React from 'react'
-import AppBar from '@mui/material/AppBar'
+import React, { useState } from 'react'
+
+// Import MaterialUI Components
 import Box from '@mui/material/Box'
 import Toolbar from '@mui/material/Toolbar'
 import Typography from '@mui/material/Typography'
 import IconButton from '@mui/material/IconButton'
-import { AccountCircle, ArrowDropDown } from '@mui/icons-material'
 import MenuItem from '@mui/material/MenuItem'
 import Menu from '@mui/material/Menu'
+import Button from '@mui/material/Button'
+import Tooltip from '@mui/material/Tooltip'
+import MuiAppBar, { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar'
+
+// Import own components
+import { DrawerHeader } from "@/components/layout/CourseDrawer";
+
+// Import styles
+import { styled } from '@mui/material/styles'
 import styles from '@/styles/Header.module.css'
+
+// Import icons
+import MenuIcon from '@mui/icons-material/Menu'
+import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown'
+import { AccountCircle } from '@mui/icons-material/'
+
+// Import images
 import Image from 'next/image'
-import { Button } from '@mui/material'
+
+// Import constants
+import { DRAWER_WIDTH, USER_SETTINGS } from '@/config/constants'
+
+// Import redux and router
 import { useAppSelector, useAppDispatch } from '@/redux/hook'
+import { setOpen } from '@/redux/features/drawerSlice'
 import { logOut } from '@/redux/features/userLoginSlice'
-import { useRouter } from 'next/navigation'
-import { USER_SETTINGS } from '@/config/constants'
+import { useRouter, useParams } from 'next/navigation'
+
+interface AppBarProps extends MuiAppBarProps {
+  open?: boolean
+}
+
+const AppBar = styled(MuiAppBar, {
+  shouldForwardProp: (prop) => prop !== 'open'
+})<AppBarProps>(({ theme, open }) => ({
+  transition: theme.transitions.create(['margin', 'width'], {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen
+  }),
+  ...(open && {
+    width: `calc(100% - ${DRAWER_WIDTH}px)`,
+    marginLeft: `${DRAWER_WIDTH}px`,
+    transition: theme.transitions.create(['margin', 'width'], {
+      easing: theme.transitions.easing.easeOut,
+      duration: theme.transitions.duration.enteringScreen
+    })
+  })
+}))
 
 export default function MainAppBar() {
+  // To determine if you are in or out of course pages
+  const params = useParams()
+  console.log(params)
+
   // Router
   const router = useRouter()
 
@@ -28,20 +73,29 @@ export default function MainAppBar() {
   const userNameState = useAppSelector(
     (state) => state.persistedReducer.userLoginState.first_name
   )
+  const open = useAppSelector(
+    (state) => state.persistedReducer.drawerState.open
+  )
 
   // Redux dispatch
   const dispatch = useAppDispatch()
 
   // Menu states
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
+  const [anchorElUser, setAnchorElUser] = useState<null | HTMLElement>(null)
 
   // Event handlers
-  const handleMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget)
+  const handleOpenUserMenu = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorElUser(event.currentTarget)
   }
 
-  const handleClose = () => {
-    setAnchorEl(null)
+  const handleCloseUserMenu = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorElUser(null)
+  }
+
+  const handleDrawerOpenClose = (
+    event: React.MouseEvent<HTMLButtonElement>
+  ) => {
+    dispatch(setOpen())
   }
 
   const handleNavigate = (rute: string) => {
@@ -55,24 +109,27 @@ export default function MainAppBar() {
 
   // Menu options for the log in user
   const logInUserOptions = (
-    <>
-      <Typography variant='h6' component='p'>
+    <Box className={styles.userBox}>
+      <Typography className={styles.userName} component='h2'>
         {userNameState}
       </Typography>
       <AccountCircle />
-      <IconButton
-        size='large'
-        aria-label='account of current user'
-        aria-controls='menu-appbar'
-        aria-haspopup='true'
-        onClick={handleMenu}
-        color='inherit'
-      >
-        <ArrowDropDown />
-      </IconButton>
+      <Tooltip title='Abrir opciones'>
+        <IconButton
+          onClick={handleOpenUserMenu}
+          size='large'
+          aria-label='account of current user'
+          aria-controls='menu-appbar'
+          aria-haspopup='true'
+          color='inherit'
+        >
+          <KeyboardArrowDownIcon />
+        </IconButton>
+      </Tooltip>
       <Menu
+        className={styles.userMenu}
         id='menu-appbar'
-        anchorEl={anchorEl}
+        anchorEl={anchorElUser}
         anchorOrigin={{
           vertical: 'top',
           horizontal: 'right'
@@ -82,11 +139,11 @@ export default function MainAppBar() {
           vertical: 'top',
           horizontal: 'right'
         }}
-        open={Boolean(anchorEl)}
-        onClose={handleClose}
+        open={Boolean(anchorElUser)}
+        onClose={handleCloseUserMenu}
       >
         {USER_SETTINGS.map((setting) => (
-          <MenuItem key={`user-opt-${setting.title}`} onClick={handleClose}>
+          <MenuItem key={`user-opt-${setting.title}`} onClick={handleCloseUserMenu}>
             <Typography
               textAlign='center'
               onClick={() => handleNavigate(setting.route)}
@@ -97,7 +154,7 @@ export default function MainAppBar() {
         ))}
         <MenuItem onClick={handleLogOut}>Cerrar sesión</MenuItem>
       </Menu>
-    </>
+    </Box>
   )
 
   // Menu options for the log out user
@@ -124,27 +181,43 @@ export default function MainAppBar() {
     </>
   )
 
+  const drawerButton = (
+    <IconButton
+      color='inherit'
+      aria-label='open drawer'
+      onClick={handleDrawerOpenClose}
+      edge='start'
+      size='large'
+    >
+      <MenuIcon />
+    </IconButton>
+  )
+
   return (
     <Box id='header'>
-      <AppBar position='static'>
+      <AppBar position='fixed' open={params.hasOwnProperty('alias') ? open : false}>
         <Toolbar className={styles.mainHeader}>
-          <div className={styles.topBarArea}>
-            <Image
-              src='/vercel.svg'
-              alt='Vercel Logo'
-              className={styles.logo}
-              width={50}
-              height={50}
-              priority
-              onClick={() => {}}
-            />
-          </div>
-
-          <div className={styles.topBarArea}>
+          <Box className={styles.topBarArea}>
+            {params.hasOwnProperty('alias') && drawerButton}
+            <Box sx={{ display: { xs: 'none', sm: 'block' } }}>
+              <Image
+                className={styles.logo}
+                hidden={open}
+                src='/vercel.svg'
+                alt='Vercel Logo'
+                width={80}
+                height={40}
+                priority
+                onClick={() => {}}
+              />
+            </Box>
+          </Box>
+          <Box className={styles.topBarArea}>
             {userLoginState ? logInUserOptions : logOutUserOptions}
-          </div>
+          </Box>
         </Toolbar>
       </AppBar>
+      <DrawerHeader />
     </Box>
   )
 }
