+++ styles/Course.module.css
@@ -3,34 +3,7 @@
 }
 
 .course > Main {
-  padding: 2rem 4rem;
-}
-
-.toolbar {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  -webkit-box-shadow: 0px 0px 15px 2px rgba(0, 0, 0, 0.35);
-  -moz-box-shadow: 0px 0px 15px 2px rgba(0, 0, 0, 0.35);
-  box-shadow: 0px 0px 15px 2px rgba(0, 0, 0, 0.35);
-}
-
-.logoBox {
-  display: flex;
-  align-items: center;
-}
-
-.userBox {
-  display: flex;
-  align-items: center;
-}
-
-.userName {
-  padding-right: 0.75rem;
-}
-
-.userMenu {
-  margin-top: 3rem;
+  padding: 2rem 5vw;
 }
 
 .moduleListItemText {
